디자인 수정

+++ intworldapp/views.py
@@ -34,9 +34,11 @@ def update(request, post_id):
     return render(request, 'update.html')
 
 def delete(request, post_id):
-    post = get_object_or_404(Post, id=post_id)
-    post.delete()
-    return redirect(reverse('index'))
+    post = Post.objects.get(id = post_id)
+    if request.method == 'POST':
+        post = get_object_or_404(Post, id=post_id)
+        post.delete()
+        return redirect(reverse('index'))
 
 
 @login_required
